requirements.txt
requests
python-dotenv
mysql-connector-python
pandas
pyarrow
lightgbm
mlflow
scikit-learn
streamlit